refactor: lift the bounded command runner out of the hwaccel probe
HardwareAccel.Probe.run_bounded/3 was the one place that ran an external
command under a hard deadline and actually killed it on expiry. The keyframe
lookup for stream-copy resumes needs the same thing, so it moves to
Mydia.BoundedCommand unchanged, tests included.
