+++ lib/mydia/streaming/hardware_accel/probe.ex
@@ -17,6 +17,7 @@ defmodule Mydia.Streaming.HardwareAccel.Probe do
 
   require Logger
 
+  alias Mydia.BoundedCommand
   alias Mydia.Streaming.HardwareAccel.Capabilities
 
   @device_glob "/dev/dri/renderD*"
@@ -133,7 +134,7 @@ defmodule Mydia.Streaming.HardwareAccel.Probe do
   end
 
   defp vainfo(device) do
-    case run_bounded("vainfo", ["--display", "drm", "--device", device]) do
+    case BoundedCommand.run("vainfo", ["--display", "drm", "--device", device], @probe_timeout_ms) do
       {:ok, output} ->
         matrix = parse_vainfo(output)
 
@@ -208,11 +209,9 @@ defmodule Mydia.Streaming.HardwareAccel.Probe do
 
     # Deliberately not Mydia.Library.Ffmpeg.run/2: that wraps System.cmd/3,
     # which blocks this process until the OS process exits with no way to
-    # bound or interrupt it. run_bounded/3 below is the same
-    # :spawn_executable + tracked-os_pid pattern FfmpegHlsTranscoder and
-    # FfmpegMp4Transcoder already use to actually kill a wedged ffmpeg
+    # bound or interrupt it. Mydia.BoundedCommand.run/3 kills a wedged ffmpeg
     # instead of merely giving up on waiting for it.
-    case run_bounded(executable(:ffmpeg_path, "ffmpeg"), args) do
+    case BoundedCommand.run(executable(:ffmpeg_path, "ffmpeg"), args, @probe_timeout_ms) do
       {:ok, _output} ->
         :ok
 
@@ -232,94 +231,5 @@ defmodule Mydia.Streaming.HardwareAccel.Probe do
   # name on PATH.
   defp executable(env_key, default_name) do
     Application.get_env(:mydia, env_key) || default_name
-  end
-
-  @doc false
-  # Runs `name` as a Port with a tracked OS pid, the same shape
-  # FfmpegHlsTranscoder.start_ffmpeg_process/1 uses, so a timeout can
-  # actually SIGKILL the child instead of merely abandoning a handle to it --
-  # Task.shutdown/2 around System.cmd/3 cannot do that, since System.cmd/3
-  # never gives the caller an OS pid to kill.
-  #
-  # Public only so the timeout and kill behaviour can be exercised directly
-  # against a generic command (`sleep`, `sh`) without depending on vainfo or
-  # ffmpeg being installed, and without a test having to wait out a real
-  # multi-second probe; nothing outside this module should call it.
-  @spec run_bounded(String.t(), [String.t()], timeout()) ::
-          {:ok, binary()} | {:error, {:exit, integer(), binary()} | :timeout | :not_found}
-  def run_bounded(name, args, timeout \\ @probe_timeout_ms) do
-    case System.find_executable(name) do
-      nil ->
-        {:error, :not_found}
-
-      path ->
-        port =
-          Port.open({:spawn_executable, path}, [
-            :binary,
-            :exit_status,
-            :stderr_to_stdout,
-            :hide,
-            args: args
-          ])
-
-        os_pid =
-          case Port.info(port, :os_pid) do
-            {:os_pid, os_pid} -> os_pid
-            nil -> nil
-          end
-
-        # An absolute deadline, not a per-receive timeout. `after` applies to
-        # each `receive`, so passing the bound straight down would restart the
-        # whole window on every {:data, _} chunk, and a child that hangs while
-        # still writing output would run indefinitely -- exactly the case this
-        # bound exists to stop.
-        deadline =
-          case timeout do
-            :infinity -> :infinity
-            ms -> System.monotonic_time(:millisecond) + ms
-          end
-
-        collect_bounded(port, os_pid, "", deadline)
-    end
-  end
-
-  defp collect_bounded(port, os_pid, buffer, deadline) do
-    receive do
-      {^port, {:data, data}} ->
-        collect_bounded(port, os_pid, buffer <> data, deadline)
-
-      {^port, {:exit_status, 0}} ->
-        {:ok, buffer}
-
-      {^port, {:exit_status, code}} ->
-        {:error, {:exit, code, buffer}}
-    after
-      remaining_ms(deadline) ->
-        kill_bounded(port, os_pid)
-        {:error, :timeout}
-    end
-  end
-
-  defp remaining_ms(:infinity), do: :infinity
-
-  defp remaining_ms(deadline),
-    do: max(deadline - System.monotonic_time(:millisecond), 0)
-
-  # Closing the port stops it relaying further messages, but does not by
-  # itself terminate a running child -- Erlang's default port behaviour on
-  # close is to keep the OS process alive when the driver was opened with
-  # :spawn_executable. SIGKILL via the same `kill -9` mechanism the
-  # transcoders use is what actually ends it, leaving no orphaned vainfo or
-  # ffmpeg process behind.
-  defp kill_bounded(port, os_pid) do
-    if os_pid, do: System.cmd("kill", ["-9", to_string(os_pid)], stderr_to_stdout: true)
-
-    try do
-      Port.close(port)
-    rescue
-      ArgumentError -> :ok
-    end
-
-    :ok
   end
 end
